+++ example/components/title.js
@@ -1,11 +1,14 @@
-const { BaseComponent, text } = require('larana-js')
+const { BaseComponent, text, layout } = require('larana-js')
 
 class TitleComponent extends BaseComponent {
 	root() {
 		const title = this.inject('title', 'Nothing')
-		// console.log(title)
 
-		return text({ value: title, style: 'h3' })
+		return layout({
+			children: [
+				text({ value: title, style: 'h2' }),
+			],
+		})
 	}
 }