+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.2.13
 
+- **fix:** tabs height
 - **feat:** provide/inject
 - **feat:** `row` and `column` components
 - **feat:** removed `lastW` and `lastH`, added `useResolution()`